@@ -17,15 +17,11 @@ class FallbackStrategies(StrategyList):
     NAME = "fallback"
 
     @staticmethod
-    def strategy_override(
-        config: deepmerge.merger.Merger, path: List, base: Any, nxt: T
-    ) -> T:
+    def strategy_override(config: deepmerge.merger.Merger, path: list, base: Any, nxt: T) -> T:
         """use nxt, and ignore base."""
         return nxt
 
     @staticmethod
-    def strategy_use_existing(
-        config: deepmerge.merger.Merger, path: List, base: T, nxt: Any
-    ) -> T:
+    def strategy_use_existing(config: deepmerge.merger.Merger, path: list, base: T, nxt: Any) -> T:
         """use base, and ignore next."""
         return base
