@@ -14,27 +14,21 @@ class SetStrategies(StrategyList):
     NAME = "set"
 
     @staticmethod
-    def strategy_union(
-        config: deepmerge.merger.Merger, path: List, base: Set, nxt: Set
-    ) -> Set:
+    def strategy_union(config: deepmerge.merger.Merger, path: list, base: set, nxt: set) -> set:
         """
         use all values in either base or nxt.
         """
         return base | nxt
 
     @staticmethod
-    def strategy_intersect(
-        config: deepmerge.merger.Merger, path: List, base: Set, nxt: Set
-    ) -> Set:
+    def strategy_intersect(config: deepmerge.merger.Merger, path: list, base: set, nxt: set) -> set:
         """
         use all values in both base and nxt.
         """
         return base & nxt
 
     @staticmethod
-    def strategy_override(
-        config: deepmerge.merger.Merger, path: List, base: Set, nxt: Set
-    ) -> Set:
+    def strategy_override(config: deepmerge.merger.Merger, path: list, base: set, nxt: set) -> set:
         """
         use the set nxt.
         """
